feat(connect): inline stellar transaction transform, deprecate @trezor/connect-plugin-stellar
Inlines the @stellar/stellar-sdk Transaction → protobuf-shape transform into
@trezor/connect's lazy-loaded stellar chunk so callers can pass a stellar-sdk
Transaction directly to TrezorConnect.stellarSignTransaction. The plugin's
10.x release becomes a deprecation stub.

Mirrors the architectural pattern from #27091 (connect-plugin-ethereum).
