--- packages/connect-plugin-stellar/__tests__/inputs.test.ts
@@ -1,13 +1,0 @@
-import * as fixtures from '../__fixtures__/inputs';
-import { transformTransaction } from '../src/index';
-
-describe('plugins/stellar', () => {
-    describe('transformTransaction', () => {
-        fixtures.transformTransactionInputs.forEach(f => {
-            it(f.description, () => {
-                const resp = transformTransaction(f.path, f.tx);
-                expect(resp).toEqual(f.result);
-            });
-        });
-    });
-});
