--- packages/connect-plugin-stellar/jest.config.js
@@ -1,5 +1,0 @@
-const baseConfig = require('../../jest.config.base.swc');
-
-module.exports = {
-    ...baseConfig,
-};
