+++ packages/connect-plugin-stellar/tsconfig.lib.json
@@ -6,9 +6,6 @@
     "include": ["./src"],
     "references": [
         {
-            "path": "../utils"
-        },
-        {
             "path": "../eslint"
         }
     ]
