+++ packages/connect-plugin-stellar/tsconfig.libESM.json
@@ -6,9 +6,6 @@
     },
     "references": [
         {
-            "path": "../utils"
-        },
-        {
             "path": "../eslint"
         }
     ]
