+++ packages/connect/package.json
@@ -122,6 +122,7 @@
         "@solana-program/token": "^0.5.1",
         "@solana-program/token-2022": "^0.4.2",
         "@solana/kit": "^2.3.0",
+        "@stellar/stellar-sdk": "14.5.0",
         "@trezor/blockchain-link": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/blockchain-link-utils": "workspace:*",
