+++ packages/connect/src/api/stellar/__tests__/transformTransaction.fixtures.ts
